requirements.txt
streamlit>=1.28.0
anthropic>=0.21.0
python-docx>=0.8.11
PyPDF2>=3.0.1
pathlib2>=2.3.7